# openk9-3.1.x/docs/openapi-yaml/datasource-openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    DocType:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createDate:
          $ref: "#/components/schemas/OffsetDateTime"
        modifiedDate:
          $ref: "#/components/schemas/OffsetDateTime"
        name:
          type: string
        description:
          type: string
...
    DocTypeField:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createDate:
          $ref: "#/components/schemas/OffsetDateTime"
        modifiedDate:
          $ref: "#/components/schemas/OffsetDateTime"
        floatBoost:
          type: number
          format: float
        keyword:
          type: boolean
        text:
          type: boolean
        numeric:
          type: boolean
        date:
          type: boolean
        boolean:
          type: boolean
        i18N:
          type: boolean
        autocomplete:
          type: boolean
        defaultExclude:
          type: boolean
        defaultBoost:
          type: boolean
        searchableAndAutocomplete:
          type: boolean
        searchableAndDate:
          type: boolean
        searchableAndText:
          type: boolean
        searchableAndI18N:
          type: boolean
        name:
          type: string
        description:
          type: string
        fieldName:
          type: string
        searchable:
          type: boolean
        sortable:
          type: boolean
        boost:
          type: number
          format: double
        fieldType:
          $ref: "#/components/schemas/FieldType"
        exclude:
          type: boolean
        jsonConfig:
          type: string
        path:
          type: string
        children:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/DocTypeField"
        docTypeFieldAndChildren:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/DocTypeField"
    DocTypeFieldDTO:
      type: object
      required:
      - name
      - searchable
      - sortable
      - fieldType
      - fieldName
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        searchable:
          type: boolean
        sortable:
          type: boolean
        boost:
          type: number
          format: double
        fieldType:
          $ref: "#/components/schemas/FieldType"
        exclude:
          type: boolean
        fieldName:
          type: string
        jsonConfig:
          type: string
...
    FieldType:
      type: string
      enum:
      - "NULL"
      - BINARY
      - BOOLEAN
      - KEYWORD
      - CONSTANT_KEYWORD
      - WILDCARD
      - DATE
      - DATE_NANOS
      - LONG
      - INTEGER
      - SHORT
      - BYTE
      - DOUBLE
      - FLOAT
      - HALF_FLOAT
      - SCALED_FLOAT
      - UNSIGNED_LONG
      - OBJECT
      - FLATTENED
      - NESTED
      - JOIN
      - LONG_RANGE
      - DOUBLE_RANGE
      - DATE_RANGE
      - IP_RANGE
      - IP
      - VERSION
      - MURMUR3
      - HISTOGRAM
      - TEXT
      - ANNOTATED_TEXT
      - COMPLETION
      - SEARCH_AS_YOU_TYPE
      - TOKEN_COUNT
      - DENSE_VECTOR
      - KNN_VECTOR
      - SPARSE_VECTOR
      - RANK_FEATURE
      - RANK_FEATURES
      - GEO_POINT
      - GEO_SHAPE
      - POINT
      - SHAPE
      - PERCOLATOR
      - I18N
...
    K9Column:
      type: string
      enum:
      - id
      - name
      - description
      - createDate
      - modifiedDate
...
    OffsetDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50-04:00
...
    PageDocTypeField:
      type: array
      items:
        $ref: "#/components/schemas/DocTypeField"
...
    Tuple2DocTypeDocTypeField:
      type: object
      properties:
        left:
          $ref: "#/components/schemas/DocType"
        right:
          $ref: "#/components/schemas/DocTypeField"
    Tuple2DocTypeLong:
      type: object
      properties:
        left:
          $ref: "#/components/schemas/DocType"
        right:
          type: integer
          format: int64
...
  securitySchemes:
    SecurityScheme:
      description: Authentication
      type: openIdConnect
paths:
...
  /api/datasource/doc-types/{id}/doc-type-fields:
    get:
      deprecated: true
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: after_id
        in: query
        schema:
          type: integer
          format: int64
          default: -1
      - name: before_id
        in: query
        schema:
          type: integer
          format: int64
          default: -1
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      - name: searchText
        in: query
        schema:
          type: string
      - name: sortBy
        in: query
        schema:
          allOf:
          - $ref: "#/components/schemas/K9Column"
          - default: createDate
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageDocTypeField"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Doc Type Fields
      tags:
      - Doc Type Resource
      security:
      - SecurityScheme:
        - k9-admin
    put:
      deprecated: true
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocTypeFieldDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tuple2DocTypeDocTypeField"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Add Doc Type Field
      tags:
      - Doc Type Resource
      security:
      - SecurityScheme:
        - k9-admin
  /api/datasource/doc-types/{id}/doc-type-fields/{docTypeFieldId}:
    delete:
      deprecated: true
      parameters:
      - name: docTypeFieldId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tuple2DocTypeLong"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Remove Doc Type Field
      tags:
      - Doc Type Resource
      security:
      - SecurityScheme:
        - k9-admin
...
info:
  title: datasource API
  version: 3.1.0